+++ wulu/src/index.js
@@ -4,6 +4,7 @@ import Economy from './build/economy';
 import {Emoticons} from './build/emoticons';
 import Mongo from './build/mongo';
 import Poll from './build/poll';
+import Sysop from './build/sysop';
 import Tour from './build/tour';
 
 export default {
@@ -13,5 +14,6 @@ export default {
   Emoticons,
   Mongo,
   Poll,
+  Sysop,
   Tour
 };