+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.6.2 / 2015-04-05
+==================
+
+  * Fix usergroups not applying to usernames with spaces
+
 1.6.1 / 2015-04-04
 ==================